requirements.txt
streamlit
python-dotenv
pyyaml
loguru
langchain
langgraph
langchain-openai
requests
graphviz
weasyprint
markdown
pydantic
PyMuPDF
tiktoken 
reportlab